+++ examples/esp32c3/examples/sw_and_hw.rs
@@ -1,6 +1,5 @@
 #![no_main]
 #![no_std]
-
 #[rtic::app(device = esp32c3, dispatchers=[FROM_CPU_INTR0, FROM_CPU_INTR1])]
 mod app {
     use esp_backtrace as _;