+++ rtic-monotonics/CHANGELOG.md
@@ -15,6 +15,11 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 - Update `esp32c3` dependency
 
+### Fixed
+
+- STM32: Make initialization more deterministic
+- STM32: Fix race condition that caused missed interrupts
+
 ## v2.0.2 - 2024-07-05
 
 ### Added