+++ rtic-monotonics/CHANGELOG.md
@@ -7,6 +7,8 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ## Unreleased
 
+## v2.0.3 - 2024-10-23
+
 ### Added
 
 - RP235x support